Event ordering preparation (#4513)
Event ordering preparation
